+++ linodecli/plugins/obj.py
@@ -101,9 +101,7 @@ def list_objects_or_buckets(get_client, args):
     """
     Lists buckets or objects
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " ls")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " ls"))
 
     parser.add_argument(
         "bucket",
@@ -170,9 +168,7 @@ def create_bucket(get_client, args):
     """
     Creates a new bucket
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " mb")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " mb"))
 
     parser.add_argument(
         "name",
@@ -194,9 +190,7 @@ def delete_bucket(get_client, args):
     """
     Deletes a bucket
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " rb")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " rb"))
 
     parser.add_argument(
         "name",
@@ -237,9 +231,7 @@ def upload_object(get_client, args):  # pylint: disable=too-many-locals
     """
     Uploads an object to object storage
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " put")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " put"))
 
     parser.add_argument(
         "file", metavar="FILE", type=str, nargs="+", help="The files to upload."
@@ -391,9 +383,7 @@ def get_object(get_client, args):
     """
     Retrieves an uploaded object and writes it to a file
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " get")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " get"))
 
     parser.add_argument(
         "bucket", metavar="BUCKET", type=str, help="The bucket the file is in."
@@ -441,9 +431,7 @@ def delete_object(get_client, args):
     """
     Removes a file from a bucket
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " del")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " del"))
 
     parser.add_argument(
         "bucket", metavar="BUCKET", type=str, help="The bucket to delete from."
@@ -478,9 +466,7 @@ def generate_url(get_client, args):
     """
     Generates a URL to an object
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " signurl")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " signurl"))
 
     parser.add_argument(
         "bucket",
@@ -534,9 +520,7 @@ def set_acl(get_client, args):
     """
     Modify Access Control List for a Bucket or Objects
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " setacl")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " setacl"))
 
     parser.add_argument(
         "bucket", metavar="BUCKET", type=str, help="The bucket to modify."
@@ -598,9 +582,7 @@ def enable_static_site(get_client, args):
     """
     Turns a bucket into a static website
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " ws-create")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " ws-create"))
 
     parser.add_argument(
         "bucket",
@@ -644,9 +626,7 @@ def static_site_info(get_client, args):
     """
     Returns info about a configured static site
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " ws-info")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " ws-info"))
 
     parser.add_argument(
         "bucket",
@@ -681,9 +661,7 @@ def show_usage(get_client, args):
     """
     Shows space used by all buckets in this cluster, and total space
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " du")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " du"))
 
     parser.add_argument(
         "bucket",
@@ -752,9 +730,7 @@ def list_all_objects(get_client, args):
     Lists all objects in all buckets
     """
     # this is for printing help when --help is in the args
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " la")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " la"))
 
     parser.parse_args(args)
 
@@ -785,9 +761,7 @@ def disable_static_site(get_client, args):
     """
     Disables static site for a bucket
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE + " du")
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE + " du"))
 
     parser.add_argument(
         "bucket",
@@ -864,9 +838,7 @@ def get_obj_args_parser():
     """
     Initialize and return the argument parser for the obj plug-in.
     """
-    parser = inherit_plugin_args(
-        ArgumentParser(PLUGIN_BASE, add_help=False)
-    )
+    parser = inherit_plugin_args(ArgumentParser(PLUGIN_BASE, add_help=False))
 
     parser.add_argument(
         "command",
