+++ linodecli/plugins/ssh.py
@@ -12,6 +12,7 @@ import argparse
 import subprocess
 import sys
 from sys import platform
+
 from linodecli.plugins import inherit_plugin_args
